refactor: rename package workManager to the worker
