+++ app/src/main/java/com/github/plplmax/notifications/data/worker/ScheduleWorkerFactory.kt
@@ -1,4 +1,4 @@
-package com.github.plplmax.notifications.data.workManager
+package com.github.plplmax.notifications.data.worker
 
 import android.content.Context
 import androidx.work.ListenableWorker
